api/pyproject.toml
[tool.poetry]
name = "mt-igcl-api"
version = "0.1.0"
description = ""
authors = ["Nicolas Huber <nicolas.huber2@uzh.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
cloudant = "^2.15.0"
mysql-connector-python = "^9.0.0"
sqlalchemy = "^2.0.35"
pymongo = "^4.10.0"
qdrant-client = "^1.11.3"
python-dotenv = "^1.0.1"
pytz = "2024.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
redis = "^5.1.1"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.12"
fastapi-pagination = "^0.12.31"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
